web-frontend/requirements.txt
appdirs==1.4.4
black==20.8b1
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dataclasses==0.8
Flask==1.1.2
google-api-core==1.26.3
google-auth==1.28.1
google-cloud==0.34.0
google-cloud-tasks==2.2.0
googleapis-common-protos==1.53.0
grpc-google-iam-v1==0.12.3
grpcio==1.37.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
libcst==0.3.18
MarkupSafe==1.1.1
mypy-extensions==0.4.3
packaging==20.9
pathspec==0.8.1
proto-plus==1.18.1
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
rsa==4.7.2
six==1.15.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.4
Werkzeug==1.0.1
